Split `SetoidTower` into multiple levels
I also made the different structures richer in term of coherence laws,
since it seemed necessary while working on Param_arrow.

I was really excited while doing the split but I'm stuggle to explain in
this commit message why exactly I wanted to do this, so... Originally, I
added these levels to be able to say things like :
 There is a morphism `(a ~ b) ~> (b ~ a)`
A setoid morphism A ~> B needs towers of height 2 on both A and B, so
the type A such that `a b: A` in the example above would need to have a
tower of height 3. It wouldn't work out if we only had a single tower
height.

The heights 0 and 1 are not needed but I figured I'd add them anyway.

The current implementation is however not ideal: the multiple level0
coexisting for each tower height confuse Rocq and it sometimes thinks
`level0`s from a tower and a coercion of the tower to a lower level are
unrelated.

Example:
```
Instance Setoid_MorphismSetoid@{i j l k | i <= l, j <= l, l <= k} `{Funext}
  {A: Type@{i}} `{Level3.SetoidTower A} {B: Type@{j}} `{Level3.SetoidTower B} :
  Setoid@{k} (A ~> B).
Proof.
  unshelve eexists.
  - exact (fun f f' => forall a a', a ~ a' ~> f a ~ f' a').
  - move=> f a a'.
    unshelve eexists.
    + move=> aR.
      apply preserves_rel.
      (* BOOM! *)
...
```